# openapi.yml
openapi: 3.1.0
info:
  title: Vehicle Rental System API
  description: API for managing vehicle rentals, customers, and vehicles
  version: 1.0.0
  contact:
    name: Vehicle Rental System
    email: support@vrs.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
paths:
...
  /api/rentals:
    get:
      tags:
        - Rentals
      summary: Get all rentals
      description: Retrieves all rentals in the system
      responses:
        '200':
          description: Rentals retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RentalDto'
    post:
      tags:
        - Rentals
      summary: Create a new rental
      description: Creates a new rental in the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentalDto'
      responses:
        '201':
          description: Rental created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalDto'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /api/rentals/{id}:
    get:
      tags:
        - Rentals
      summary: Get rental by ID
      description: Retrieves a specific rental by its ID
      parameters:
        - name: id
          in: path
          description: Rental ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Rental found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalDto'
        '404':
          description: Rental not found
    put:
      tags:
        - Rentals
      summary: Update rental
      description: Updates an existing rental
      parameters:
        - name: id
          in: path
          description: Rental ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentalDto'
      responses:
        '200':
          description: Rental updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalDto'
        '404':
          description: Rental not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
    delete:
      tags:
        - Rentals
      summary: Delete rental
      description: Deletes a rental from the system
      parameters:
        - name: id
          in: path
          description: Rental ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Rental deleted successfully
        '404':
          description: Rental not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /api/rentals/{id}/status:
    patch:
      tags:
        - Rentals
      summary: Update rental status
      description: Updates the status of an existing rental
      parameters:
        - name: id
          in: path
          description: Rental ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [PENDING, ACTIVE, COMPLETED, CANCELLED]
              required:
                - status
      responses:
        '200':
          description: Rental status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalDto'
        '404':
          description: Rental not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
components:
  schemas:
...
    RentalDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Rental ID
        vehicleId:
          type: integer
          format: int64
          description: ID of the rented vehicle
        vehicleDetails:
          type: string
          description: Details of the rented vehicle
        customerId:
          type: integer
          format: int64
          description: ID of the customer
        customerName:
          type: string
          description: Name of the customer
        startDate:
          type: string
          format: date
          description: Rental start date
        endDate:
          type: string
          format: date
          description: Rental end date
        actualReturnDate:
          type: string
          format: date
          description: Actual return date
        status:
          $ref: '#/components/schemas/RentalStatus'
        totalCost:
          type: number
          format: decimal
          description: Total rental cost
        notes:
          type: string
          description: Additional notes
    ErrorResponseDto:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Error type
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error
        path:
          type: string
          description: Request path that caused the error
...
    RentalStatus:
      type: string
      enum:
        - PENDING
        - ACTIVE
        - COMPLETED
        - CANCELLED
      description: Current status of the rental
...